Let token-age filters express seconds, as a fraction of a minute (#30)
The age fields are labelled "minutes", and people filtering the very freshest
launches think in seconds. 0.25 is how you say fifteen of them without
inventing a second unit - but nothing in the stack would have honoured it.

Three separate places had to change, and fixing fewer than all three would have
produced a control that looks like it works and doesn't:

  - The API rejected it outright. `z.number().int()` on both bounds, so 0.25
    was a validation error rather than a quarter of a minute.
  - Postgres would have truncated it. UserFilter's bounds were Int; now Float.
  - ...and even stored perfectly, it would have matched nothing, because
    ageMinutes itself was rounded to the nearest whole minute at capture. A
    token 15 seconds old recorded as 0 and one 40 seconds old as 1, so no
    fractional bound could ever tell them apart. Now kept to two decimals
    (0.01 min = 0.6s), which is finer than the scan cadence can resolve and
    still readable in a log.

TokenSnapshot.ageMinutes widens Int -> Float alongside it. Both migrations are
plain widenings (ALTER COLUMN ... SET DATA TYPE DOUBLE PRECISION), so existing
whole-minute rows carry over untouched.

Worth being straight about the limit this does NOT remove: the scanner runs
once a minute, so a token is only ever evaluated at scan time. A sub-minute
*minimum* age is therefore approximate in practice - it is the stored age that
is now precise, not the sampling.

Verified: 12 enrich tests including two new ones that pin the exact behaviour
the old rounding destroyed - a 15-second-old token reads 0.25 rather than 0,
and 15s vs 40s no longer collapse to the same value. Full suite green
(140 + 55 + 48).

+++ apps/api/src/routes/filters.ts
@@ -17,8 +17,10 @@ function buildFilterInputSchema(env: Env) {
     maxDevWalletPct: z.number().min(0).max(100).nullable().optional(),
     maxRiskScore: z.number().min(0).max(100).nullable().optional(),
     excludeCriticalRiskFlags: z.boolean().default(false),
-    minTokenAgeMinutes: z.number().int().nonnegative().nullable().optional(),
-    maxTokenAgeMinutes: z.number().int().nonnegative().nullable().optional(),
+    // Not .int(): a fraction of a minute is how you express seconds here, and rejecting 0.25
+    // would make the field's own placeholder a lie.
+    minTokenAgeMinutes: z.number().nonnegative().nullable().optional(),
+    maxTokenAgeMinutes: z.number().nonnegative().nullable().optional(),
     narrativeKeywords: z.array(z.string()).default([]),
     minScore: z.number().min(0).max(100).nullable().optional(),
     maxFreshTop10WalletPct: z.number().min(0).max(100).nullable().optional(),
